+++ spring-mvc/src/main/java/db/PlantInterface.java
@@ -5,7 +5,7 @@ import model.Plant;
 import java.util.List;
 
 public interface PlantInterface {
-    public Plant getPlant(int id);
+    public Plant getPlant(long id);
     public void addPlant(Plant plant);
     public List<Plant> getAllePlanten();
     public void verwijderPlant(int id);
